% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{emmHelper}
\alias{emmHelper}
\title{emmHelper}
\usage{
emmHelper(data, newdata, ck)
}
\arguments{
\item{data}{data is a data.frame of model data}

\item{newdata}{newdata is a data.frame of newly generated data}

\item{ck}{ck is a boll vector of expected specs in mynames}
}
\value{
a list of emmeams, se and confidence interval of each expected spec
}
\description{
function to calculate emm, se and confidence intervals
}
