% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myEMMean.R
\name{myEMM}
\alias{myEMM}
\title{my estimated marginal mean calculator}
\usage{
myEMM(object, specs)
}
\arguments{
\item{object}{object is a linear regression model}

\item{specs}{a variable or a vector of variables to calculate EMM with}
}
\value{
a list of EMMs for each spec
}
\description{
Obtain estimated marginal mean and its confidence interval based on given linear models.
}
\details{
This is some description of this function.
}
\examples{
myEMM(lm(mpg ~ disp + wt + factor(cyl), data = mtcars), c('wt', 'cyl'))
}
