% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{myRefGrid}
\alias{myRefGrid}
\title{myRefGrid}
\usage{
myRefGrid(data, classes)
}
\arguments{
\item{data}{data is a data.frame of model data}

\item{classes}{classes is a vector of classses of variables in the data}
}
\value{
a list of reference grid of each variable
}
\description{
function to build reference grid of each variable
}
