% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{nameHelper}
\alias{nameHelper}
\title{nameHelper}
\usage{
nameHelper(names)
}
\arguments{
\item{names}{names is the original vector of names}
}
\value{
a vector of EMMs for each specs
}
\description{
name function to remove factor() and I()
}
