% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{specs_ck}
\alias{specs_ck}
\title{specs_ck}
\usage{
specs_ck(specs, mynames)
}
\arguments{
\item{specs}{specs is a vector of expected specs}

\item{mynames}{mynames is a vector of names}
}
\value{
a boll vector of expected specs in mynames or warning when there is specs not in mynames
}
\description{
function to check specs to calculate included or not
}
