% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myEMMean.R
\name{myEMM}
\alias{myEMM}
\title{my estimated marginal mean calculator}
\usage{
myEMM(object, specs)
}
\arguments{
\item{object}{object is a lm model}

\item{specs}{a vector or variable to calculate EMM with}
}
\value{
a list of EMMs for each specs
}
\description{
the function is written for biostat625 HW4
}
\details{
This is some description of this function.

you can use this function to calculate estimated marginal mean and its confidence interval.
}
\examples{
myEMM(model1, 'Gender')
}
