% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHelpers.R
\name{xGen}
\alias{xGen}
\title{xGen}
\usage{
xGen(ref_grid, data, classes)
}
\arguments{
\item{ref_grid}{ref_grid is a list of reference grid of each variable}

\item{data}{data is a data.frame of model data}

\item{classes}{classes is a vector of classses of variables in the data}
}
\value{
a data.frame of newly generated data
}
\description{
function to generate new X according to each grid
}
